+++ cmd/apex/deploy/deploy.go
@@ -7,7 +7,6 @@ import (
 	"github.com/tj/cobra"
 
 	"github.com/apex/apex/cmd/apex/root"
-	"github.com/apex/apex/stats"
 	"github.com/apex/apex/utils"
 )
 
@@ -23,6 +22,9 @@ var concurrency int
 // alias.
 var alias string
 
+// zip path.
+var zip string
+
 // example output.
 const example = `
     Deploy all functions
@@ -36,6 +38,9 @@ const example = `
 
     Deploy functions in a different project
     $ apex deploy -C ~/dev/myapp
+
+    Deploy function with existing zip
+    $ apex build > out.zip && apex deploy foo --zip out.zip
 
     Deploy all functions starting with "auth"
     $ apex deploy auth*`
@@ -56,32 +61,18 @@ func init() {
 	f.StringSliceVarP(&env, "set", "s", nil, "Set environment variable")
 	f.StringVarP(&envFile, "env-file", "E", "", "Set environment variables from JSON file")
 	f.StringVarP(&alias, "alias", "a", "current", "Function alias")
+	f.StringVarP(&zip, "zip", "z", "", "Zip path")
 	f.IntVarP(&concurrency, "concurrency", "c", 5, "Concurrent deploys")
 }
 
 // Run command.
 func run(c *cobra.Command, args []string) error {
-	stats.Track("Deploy", map[string]interface{}{
-		"concurrency": concurrency,
-		"has_alias":   alias != "",
-		"env":         len(env),
-		"args":        len(args),
-	})
-
 	root.Project.Concurrency = concurrency
 	root.Project.Alias = alias
+	root.Project.Zip = zip
 
 	if err := root.Project.LoadFunctions(args...); err != nil {
 		return err
-	}
-
-	for _, fn := range root.Project.Functions {
-		stats.Track("Deploy Function", map[string]interface{}{
-			"runtime":      fn.Runtime,
-			"has_alias":    alias != "",
-			"has_env_file": envFile != "",
-			"env":          len(env),
-		})
 	}
 
 	if envFile != "" {