+++ plugins/golang/golang.go
@@ -2,6 +2,8 @@
 package golang
 
 import (
+	"strings"
+
 	"github.com/apex/apex/function"
 	"github.com/apex/apex/plugins/nodejs"
 )
@@ -11,7 +13,7 @@ func init() {
 }
 
 const (
-	// Runtime name used by Apex
+	// Runtime for inference.
 	Runtime = "golang"
 )
 
@@ -20,7 +22,7 @@ type Plugin struct{}
 
 // Open adds the shim and golang defaults.
 func (p *Plugin) Open(fn *function.Function) error {
-	if fn.Runtime != Runtime {
+	if !strings.HasPrefix(fn.Runtime, "golang") {
 		return nil
 	}
 
@@ -29,7 +31,11 @@ func (p *Plugin) Open(fn *function.Function) error {
 	}
 
 	fn.Shim = true
-	fn.Runtime = nodejs.Runtime43
+	fn.Runtime = nodejs.Runtime
+
+	if fn.Handler == "" {
+		fn.Handler = "index.handle"
+	}
 
 	if fn.Hooks.Clean == "" {
 		fn.Hooks.Clean = "rm -f main"