+++ project/project.go
@@ -4,7 +4,6 @@ package project
 import (
 	"bytes"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -14,6 +13,7 @@ import (
 	"github.com/apex/log"
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/lambda/lambdaiface"
+	"github.com/pkg/errors"
 	"github.com/tj/go-sync/semaphore"
 	"gopkg.in/validator.v2"
 
@@ -33,7 +33,7 @@ const (
 	DefaultTimeout = 3
 
 	// DefaultRetainedVersions defines numbers of retained versions
-	DefaultRetainedVersions = 10
+	DefaultRetainedVersions = 25
 
 	// functions directory
 	functionsDir = "functions"
@@ -55,6 +55,7 @@ type Config struct {
 	Environment        map[string]string `json:"environment"`
 	Hooks              hooks.Hooks       `json:"hooks"`
 	VPC                vpc.VPC           `json:"vpc"`
+	Zip                string            `json:"zip"`
 }
 
 // Project represents zero or more Lambda functions.
@@ -154,7 +155,7 @@ func (p *Project) LoadFunctions(patterns ...string) error {
 	for _, name := range names {
 		match, err := matches(name, patterns)
 		if err != nil {
-			return err
+			return errors.Wrapf(err, "matching %s", name)
 		}
 
 		if !match {
@@ -163,7 +164,7 @@ func (p *Project) LoadFunctions(patterns ...string) error {
 
 		fn, err := p.LoadFunction(name)
 		if err != nil {
-			return err
+			return errors.Wrapf(err, "loading %s", name)
 		}
 
 		p.Functions = append(p.Functions, fn)
@@ -348,6 +349,7 @@ func (p *Project) LoadFunctionByPath(name, path string) (*function.Function, err
 			Environment:      copyStringMap(p.Config.Environment),
 			RetainedVersions: p.RetainedVersions,
 			VPC:              copyVPC(p.VPC),
+			Zip:              p.Zip,
 		},
 		Name:       name,
 		Path:       path,